Correcting compiler conversion warnings

+++ IGraphics/Controls/IVMeterControl.h
@@ -152,7 +152,7 @@ class IVLEDMeterControl : public IVMeterControl<MAXNC>
 	{
 		const int totalNSegs = IVMeterControl<MAXNC>::mNSteps;
 		const EDirection dir = IVMeterControl<MAXNC>::mDirection;
-		const float val      = IVMeterControl<MAXNC>::GetValue(chIdx);
+		const float val      = static_cast<float>(IVMeterControl<MAXNC>::GetValue(chIdx));
 		const float valPos   = (dir == EDirection::Vertical)
 								 ? r.B - (val * r.H())
 								 : r.R - ((1.f - val) * r.W());  // threshold position for testing segment
